+++ CHANGELOG.md
@@ -3,12 +3,12 @@
  * @Author: lkxian
  * @@@Email: lkxian888@163.com
  * @Date: 2022-09-27 10:50:31
- * @LastEditTime: 2022-09-27 10:56:13
+ * @LastEditTime: 2022-09-27 14:47:15
 -->
 
 # Change Log
 
-All notable changes to the "lkx-web-dev-extension" extension will be documented in this file.
+All notable changes to the "webdevplugin.lkx" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.